Доработки для продакшена

+++ src/main/java/ru/rushydro/vniig/entry/Sensor.java
@@ -56,7 +56,7 @@ public class Sensor {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
